Fix: multiple token copies of Aurelia didn't trigger additional combat
Cause: the checkTrigger was trying to find the source as a card but if the source was a token, getCard wouldn't find anything, causing sourceCard to be null. The first token copy would trigger, but additional copies attacking for the first time were using the same counter in game.getState because sourceCard was null. By using getPermanent, checkTrigger can now find the token that was the source of the trigger and can accurately track amountAttacks for each Aurelia permanent.