CubeCobraDownloader.py Utility script for adding cubes
A python script to automatically download cubes from CubeCobra and place the list of cards in the cube folder with automatic naming. It also instructs you how to edit the 3 config files to implement it fully.
Tested by quering 2 different cubes and running client server locally and drafting a full drafting. Hopefully it will be usefull.